Separated noether_gui classes into individual files; updated for changes to noether_tpp
