+++ noether_gui/include/noether_gui/widgets/tool_path_modifiers/direction_of_travel_orientation_modifier_widget.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#include <noether_gui/widgets.h>
+
+#include <noether_tpp/core/tool_path_modifier.h>
+
+namespace noether
+{
+struct DirectionOfTravelOrientationModifierWidget : public ToolPathModifierWidget
+{
+  Q_OBJECT
+  using ToolPathModifierWidget::ToolPathModifierWidget;
+  ToolPathModifier::ConstPtr create() const override;
+};
+
+}  // namespace noether
