+++ noether_gui/include/noether_gui/widgets/tool_path_modifiers/snake_organization_modifier_widget.h
@@ -1,0 +1,18 @@
+#pragma once
+
+#include <noether_gui/widgets.h>
+
+#include <noether_tpp/core/tool_path_modifier.h>
+
+namespace noether
+{
+struct SnakeOrganizationModifierWidget : public ToolPathModifierWidget
+{
+  Q_OBJECT
+public:
+  using ToolPathModifierWidget::ToolPathModifierWidget;
+
+  ToolPathModifier::ConstPtr create() const override;
+};
+
+}  // namespace noether
