+++ noether_gui/include/noether_gui/widgets/tool_path_modifiers/uniform_orientation_modifier_widget.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#include <noether_gui/widgets.h>
+
+#include <noether_tpp/core/tool_path_modifier.h>
+
+namespace noether
+{
+struct UniformOrientationModifierWidget : public ToolPathModifierWidget
+{
+  Q_OBJECT
+  using ToolPathModifierWidget::ToolPathModifierWidget;
+  ToolPathModifier::ConstPtr create() const override;
+};
+
+}  // namespace noether
