+++ noether_gui/include/noether_gui/widgets/tool_path_planners/raster/direction_generators/principal_axis_direction_generator_widget.h
@@ -8,28 +8,8 @@ class QDoubleSpinBox;
 class QFormLayout;
 class QLabel;
 
-namespace Ui
-{
-class Vector3dEditor;
-}
-
 namespace noether
 {
-class FixedDirectionGeneratorWidget : public DirectionGeneratorWidget
-{
-  Q_OBJECT
-public:
-  FixedDirectionGeneratorWidget(QWidget* parent = nullptr);
-
-  DirectionGenerator::ConstPtr create() const override;
-
-  void configure(const YAML::Node&) override;
-  void save(YAML::Node&) const override;
-
-private:
-  Ui::Vector3dEditor* ui_;
-};
-
 class PrincipalAxisDirectionGeneratorWidget : public DirectionGeneratorWidget
 {
   Q_OBJECT
