+++ noether_gui/include/noether_gui/widgets/tool_path_planners/raster/origin_generators/fixed_origin_generator_widget.h
@@ -26,22 +26,4 @@ private:
   Ui::Vector3dEditor* ui_;
 };
 
-struct CentroidOriginGeneratorWidget : public OriginGeneratorWidget
-{
-  Q_OBJECT
-public:
-  using BaseWidget::BaseWidget;
-
-  OriginGenerator::ConstPtr create() const override;
-};
-
-struct AABBOriginGeneratorWidget : public OriginGeneratorWidget
-{
-  Q_OBJECT
-public:
-  using BaseWidget::BaseWidget;
-
-  OriginGenerator::ConstPtr create() const override;
-};
-
 }  // namespace noether
