+++ noether_gui/src/widgets/tool_path_planners/raster/raster_planner_widget.cpp
@@ -1,4 +1,4 @@
-#include <noether_gui/widgets/raster_planner_widget.h>
+#include <noether_gui/widgets/tool_path_planners/raster/raster_planner_widget.h>
 #include "ui_raster_planner_widget.h"
 #include <noether_gui/plugin_interface.h>
 #include <noether_gui/widgets/collapsible_area_widget.h>
